refactor(ia): created plugin specific action card
forces reuse between analytics and plugins sections
